refactor: move chart component into components
