+++ frontend/src/pages/org-admin/statistic/statistic-users.js
@@ -8,7 +8,7 @@ import { orgAdminAPI } from '../../../utils/org-admin-api';
 import Loading from '../../../components/loading';
 import { Utils } from '../../../utils/utils';
 import toaster from '../../../components/toast';
-import Chart from '../../../chart';
+import Chart from '../../../components/chart';
 
 const OrgStatisticUsers = () => {
   const [isLoading, setLoading] = useState(true);
