+++ frontend/src/pages/sys-admin/statistic/statistic-file.js
@@ -6,7 +6,7 @@ import Loading from '../../../components/loading';
 import { gettext } from '../../../utils/constants';
 import { Utils } from '../../../utils/utils';
 import toaster from '../../../components/toast';
-import Chart from '../../../chart';
+import Chart from '../../../components/chart';
 
 import '../../../css/system-stat.css';
