+++ frontend/src/pages/sys-admin/statistic/statistic-traffic.js
@@ -8,7 +8,7 @@ import OrgsTraffic from './statistic-traffic-orgs';
 import UsersTraffic from './statistic-traffic-users';
 import { Utils } from '../../../utils/utils';
 import toaster from '../../../components/toast';
-import Chart from '../../../chart';
+import Chart from '../../../components/chart';
 
 class StatisticTraffic extends React.Component {
